the way how native libraries can be loaded on Windows does not work ...
on Linux. It works when I add the library folder via

```
export LD_LIBRARY_PATH=/path/to/bin/folder

## and then the Maven test works:
mvn test
```

like in this answer: https://stackoverflow.com/a/7284911

So I need to figure out a platform indipendent way to load
these libs...

At least the test is running which means that the compilation
works. However, the lib-names that should be passed into
the #[link] attribute are different:

* libumfpack on Windows
* umfpack on Linux

-> this needs to be also fixed; probably via the build.rs script