we need to include a compilation condition for
including UMFPACK links or not. Fortunatelly,
this is quite easy in Rust: we just have to
pass a `--cfg umfpack` flag to the compiler
via the `RUSTFLAGS` variable. We could add
something like `withumf` to our final library
name when UMFPACK bindings are included.

When loading the library from Java, we first
would try to find the `withumf` version and
then load the respecive libraries (with
dependencies).